performance_analysis/requirements.txt
pandas
numpy
matplotlib
seaborn
scipy
pillow